scripts/resume_mapping/requirements.txt
pypdf
unidecode
spacy
openai
python-dotenv